(AUTOMATIC) update 2_1_3's installer dumps from contrexxlabs_contrexx_2_1_3
git-svn-id: http://svn.contrexxlabs.com/contrexx/branches/contrexx_2_1@7117 5ce06e5d-2b28-0410-aa6a-b3cd9cf52d8a

+++ installer/data/contrexx_dump_structure.sql
@@ -1636,7 +1636,7 @@ SET character_set_client = utf8;
 CREATE TABLE `contrexx_module_gallery_categories` (
   `id` int(11) NOT NULL auto_increment,
   `pid` int(11) NOT NULL default '0',
-  `sorting` tinyint(3) NOT NULL default '0',
+  `sorting` int(6) NOT NULL default '0',
   `status` set('0','1') NOT NULL default '1',
   `comment` set('0','1') NOT NULL default '0',
   `voting` set('0','1') NOT NULL default '0',
@@ -1691,7 +1691,7 @@ CREATE TABLE `contrexx_module_gallery_pictures` (
   `validated` set('0','1') NOT NULL default '0',
   `status` set('0','1') NOT NULL default '1',
   `catimg` set('0','1') NOT NULL default '0',
-  `sorting` smallint(3) unsigned NOT NULL default '999',
+  `sorting` int(6) unsigned NOT NULL default '999',
   `size_show` set('0','1') NOT NULL default '1',
   `path` text NOT NULL,
   `link` text NOT NULL,
